:zap: sorted files, extracted loadOptions
