+++ packages/nodes-base/nodes/Notion/NotionTrigger.node.ts
@@ -7,9 +7,9 @@ import type {
 } from 'n8n-workflow';
 
 import moment from 'moment-timezone';
-import { notionApiRequest, simplifyObjects } from './GenericFunctions';
+import { notionApiRequest, simplifyObjects } from './shared/GenericFunctions';
 
-import { getDatabases } from './SearchFunctions';
+import { listSearch } from './shared/methods';
 
 export class NotionTrigger implements INodeType {
 	description: INodeTypeDescription = {
@@ -142,9 +142,7 @@ export class NotionTrigger implements INodeType {
 	};
 
 	methods = {
-		listSearch: {
-			getDatabases,
-		},
+		listSearch,
 	};
 
 	async poll(this: IPollFunctions): Promise<INodeExecutionData[][] | null> {
