+++ packages/nodes-base/nodes/Notion/shared/descriptions/DatabasePageDescription.ts
@@ -1,6 +1,6 @@
 import type { INodeProperties } from 'n8n-workflow';
 
-import { getConditions, getSearchFilters } from './GenericFunctions';
+import { getConditions, getSearchFilters } from '../GenericFunctions';
 
 import { blocks, text } from './Blocks';
