+++ packages/nodes-base/nodes/Notion/v1/NotionV1.node.ts
@@ -10,23 +10,23 @@ import type {
 } from 'n8n-workflow';
 
 import moment from 'moment-timezone';
-import type { SortData } from '../GenericFunctions';
+import type { SortData } from '../shared/GenericFunctions';
 import {
 	extractDatabaseId,
 	extractDatabaseMentionRLC,
 	extractPageId,
 	formatBlocks,
 	formatTitle,
-	getBlockTypes,
+	getBlockTypesOptions,
 	mapFilters,
 	mapProperties,
 	mapSorting,
 	notionApiRequest,
 	notionApiRequestAllItems,
 	simplifyObjects,
-} from '../GenericFunctions';
+} from '../shared/GenericFunctions';
 
-import { getDatabases } from '../SearchFunctions';
+import { listSearch } from '../shared/methods';
 import { versionDescription } from './VersionDescription';
 
 export class NotionV1 implements INodeType {
@@ -40,9 +40,7 @@ export class NotionV1 implements INodeType {
 	}
 
 	methods = {
-		listSearch: {
-			getDatabases,
-		},
+		listSearch,
 		loadOptions: {
 			async getDatabaseProperties(this: ILoadOptionsFunctions): Promise<INodePropertyOptions[]> {
 				const returnData: INodePropertyOptions[] = [];
@@ -104,7 +102,7 @@ export class NotionV1 implements INodeType {
 				return returnData;
 			},
 			async getBlockTypes(this: ILoadOptionsFunctions): Promise<INodePropertyOptions[]> {
-				return getBlockTypes();
+				return getBlockTypesOptions();
 			},
 			async getPropertySelectValues(this: ILoadOptionsFunctions): Promise<INodePropertyOptions[]> {
 				const [name, type] = (this.getCurrentNodeParameter('&key') as string).split('|');
