+++ packages/nodes-base/nodes/Notion/v2/NotionV2.node.ts
@@ -1,17 +1,14 @@
 import type {
 	IExecuteFunctions,
 	IDataObject,
-	ILoadOptionsFunctions,
 	INodeExecutionData,
-	INodePropertyOptions,
 	INodeType,
 	INodeTypeBaseDescription,
 	INodeTypeDescription,
 } from 'n8n-workflow';
 import { jsonParse, NodeApiError } from 'n8n-workflow';
 
-import moment from 'moment-timezone';
-import type { SortData, FileRecord } from '../GenericFunctions';
+import type { SortData, FileRecord } from '../shared/GenericFunctions';
 import {
 	downloadFiles,
 	extractDatabaseId,
@@ -19,7 +16,6 @@ import {
 	extractPageId,
 	formatBlocks,
 	formatTitle,
-	getBlockTypes,
 	mapFilters,
 	mapProperties,
 	mapSorting,
@@ -29,9 +25,10 @@ import {
 	simplifyBlocksOutput,
 	simplifyObjects,
 	validateJSON,
-} from '../GenericFunctions';
+} from '../shared/GenericFunctions';
 
-import { getDatabases } from '../SearchFunctions';
+import { listSearch } from '../shared/methods';
+import { loadOptions } from './methods';
 import { versionDescription } from './VersionDescription';
 
 export class NotionV2 implements INodeType {
@@ -44,191 +41,7 @@ export class NotionV2 implements INodeType {
 		};
 	}
 
-	methods = {
-		listSearch: {
-			getDatabases,
-		},
-		loadOptions: {
-			async getDatabaseProperties(this: ILoadOptionsFunctions): Promise<INodePropertyOptions[]> {
-				const returnData: INodePropertyOptions[] = [];
-				const databaseId = this.getCurrentNodeParameter('databaseId', {
-					extractValue: true,
-				}) as string;
-				const { properties } = await notionApiRequest.call(this, 'GET', `/databases/${databaseId}`);
-				for (const key of Object.keys(properties as IDataObject)) {
-					//remove parameters that cannot be set from the API.
-					if (
-						![
-							'created_time',
-							'last_edited_time',
-							'created_by',
-							'last_edited_by',
-							'formula',
-							'rollup',
-						].includes(properties[key].type as string)
-					) {
-						returnData.push({
-							name: `${key}`,
-							value: `${key}|${properties[key].type}`,
-						});
-					}
-				}
-				returnData.sort((a, b) => {
-					if (a.name.toLocaleLowerCase() < b.name.toLocaleLowerCase()) {
-						return -1;
-					}
-					if (a.name.toLocaleLowerCase() > b.name.toLocaleLowerCase()) {
-						return 1;
-					}
-					return 0;
-				});
-				return returnData;
-			},
-			async getFilterProperties(this: ILoadOptionsFunctions): Promise<INodePropertyOptions[]> {
-				const returnData: INodePropertyOptions[] = [];
-				const databaseId = this.getCurrentNodeParameter('databaseId', {
-					extractValue: true,
-				}) as string;
-				const { properties } = await notionApiRequest.call(this, 'GET', `/databases/${databaseId}`);
-				for (const key of Object.keys(properties as IDataObject)) {
-					returnData.push({
-						name: `${key}`,
-						value: `${key}|${properties[key].type}`,
-					});
-				}
-				returnData.sort((a, b) => {
-					if (a.name.toLocaleLowerCase() < b.name.toLocaleLowerCase()) {
-						return -1;
-					}
-					if (a.name.toLocaleLowerCase() > b.name.toLocaleLowerCase()) {
-						return 1;
-					}
-					return 0;
-				});
-				return returnData;
-			},
-			async getBlockTypes(this: ILoadOptionsFunctions): Promise<INodePropertyOptions[]> {
-				return getBlockTypes();
-			},
-			async getPropertySelectValues(this: ILoadOptionsFunctions): Promise<INodePropertyOptions[]> {
-				const [name, type] = (this.getCurrentNodeParameter('&key') as string).split('|');
-				const databaseId = this.getCurrentNodeParameter('databaseId', {
-					extractValue: true,
-				}) as string;
-				const resource = this.getCurrentNodeParameter('resource') as string;
-				const operation = this.getCurrentNodeParameter('operation') as string;
-				const { properties } = await notionApiRequest.call(this, 'GET', `/databases/${databaseId}`);
-				if (resource === 'databasePage') {
-					if (['multi_select', 'select', 'status'].includes(type) && operation === 'getAll') {
-						return properties[name][type].options.map((option: IDataObject) => ({
-							name: option.name,
-							value: option.name,
-						}));
-					} else if (
-						['multi_select', 'select', 'status'].includes(type) &&
-						['create', 'update'].includes(operation)
-					) {
-						return properties[name][type].options.map((option: IDataObject) => ({
-							name: option.name,
-							value: option.name,
-						}));
-					}
-				}
-				return properties[name][type].options.map((option: IDataObject) => ({
-					name: option.name,
-					value: option.id,
-				}));
-			},
-			async getUsers(this: ILoadOptionsFunctions): Promise<INodePropertyOptions[]> {
-				const returnData: INodePropertyOptions[] = [];
-				const users = await notionApiRequestAllItems.call(this, 'results', 'GET', '/users');
-				for (const user of users) {
-					if (user.type === 'person') {
-						returnData.push({
-							name: user.name,
-							value: user.id,
-						});
-					}
-				}
-				return returnData;
-			},
-			async getDatabaseIdFromPage(this: ILoadOptionsFunctions): Promise<INodePropertyOptions[]> {
-				const returnData: INodePropertyOptions[] = [];
-				const pageId = extractPageId(
-					this.getCurrentNodeParameter('pageId', { extractValue: true }) as string,
-				);
-				const {
-					parent: { database_id: databaseId },
-				} = await notionApiRequest.call(this, 'GET', `/pages/${pageId}`);
-				const { properties } = await notionApiRequest.call(this, 'GET', `/databases/${databaseId}`);
-				for (const key of Object.keys(properties as IDataObject)) {
-					//remove parameters that cannot be set from the API.
-					if (
-						![
-							'created_time',
-							'last_edited_time',
-							'created_by',
-							'last_edited_by',
-							'formula',
-							'rollup',
-						].includes(properties[key].type as string)
-					) {
-						returnData.push({
-							name: `${key}`,
-							value: `${key}|${properties[key].type}`,
-						});
-					}
-				}
-				returnData.sort((a, b) => {
-					if (a.name.toLocaleLowerCase() < b.name.toLocaleLowerCase()) {
-						return -1;
-					}
-					if (a.name.toLocaleLowerCase() > b.name.toLocaleLowerCase()) {
-						return 1;
-					}
-					return 0;
-				});
-				return returnData;
-			},
-
-			async getDatabaseOptionsFromPage(
-				this: ILoadOptionsFunctions,
-			): Promise<INodePropertyOptions[]> {
-				const pageId = extractPageId(
-					this.getCurrentNodeParameter('pageId', { extractValue: true }) as string,
-				);
-				const [name, type] = (this.getCurrentNodeParameter('&key') as string).split('|');
-				const {
-					parent: { database_id: databaseId },
-				} = await notionApiRequest.call(this, 'GET', `/pages/${pageId}`);
-				const { properties } = await notionApiRequest.call(this, 'GET', `/databases/${databaseId}`);
-				return properties[name][type].options.map((option: IDataObject) => ({
-					name: option.name,
-					value: option.name,
-				}));
-			},
-
-			// Get all the timezones to display them to user so that they can
-			// select them easily
-			async getTimezones(this: ILoadOptionsFunctions): Promise<INodePropertyOptions[]> {
-				const returnData: INodePropertyOptions[] = [];
-				for (const timezone of moment.tz.names()) {
-					const timezoneName = timezone;
-					const timezoneId = timezone;
-					returnData.push({
-						name: timezoneName,
-						value: timezoneId,
-					});
-				}
-				returnData.unshift({
-					name: 'Default',
-					value: 'default',
-					description: 'Timezone set in n8n',
-				});
-				return returnData;
-			},
-		},
-	};
+	methods = { listSearch, loadOptions };
 
 	async execute(this: IExecuteFunctions): Promise<INodeExecutionData[][]> {
 		const items = this.getInputData();
