+++ packages/nodes-base/nodes/Notion/v2/VersionDescription.ts
@@ -1,14 +1,17 @@
 /* eslint-disable n8n-nodes-base/node-filename-against-convention */
 import type { INodeTypeDescription } from 'n8n-workflow';
-import { databaseFields, databaseOperations } from '../DatabaseDescription';
+import { databaseFields, databaseOperations } from '../shared/descriptions/DatabaseDescription';
 
-import { userFields, userOperations } from '../UserDescription';
+import { userFields, userOperations } from '../shared/descriptions/UserDescription';
 
-import { pageFields, pageOperations } from '../PageDescription';
+import { pageFields, pageOperations } from '../shared/descriptions/PageDescription';
 
-import { blockFields, blockOperations } from '../BlockDescription';
+import { blockFields, blockOperations } from '../shared/descriptions/BlockDescription';
 
-import { databasePageFields, databasePageOperations } from '../DatabasePageDescription';
+import {
+	databasePageFields,
+	databasePageOperations,
+} from '../shared/descriptions/DatabasePageDescription';
 
 export const versionDescription: INodeTypeDescription = {
 	displayName: 'Notion',
