+++ packages/nodes-base/nodes/Notion/v2/methods/index.ts
@@ -1,0 +1,1 @@
+export * as loadOptions from './loadOptions';
